% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_conversion.R
\name{a_to_q}
\alias{a_to_q}
\title{Conversion from annual to quarterly}
\usage{
a_to_q(x, type)
}
\arguments{
\item{type}{"sum", "average", "first" or "last".}
}
\description{
Based on the td function in the tempdisagg package. The Denton-Cholette method is in the
documentation as a good default method if you don't have an indicator series.
The type of conversion, or type, needs to be defined. It can be used with
conversion "sum", "average", "first" or "last".
}
\details{
This is used in the a_to_q_xts function.
}

