% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_conversion.R
\name{a_to_q_xts}
\alias{a_to_q_xts}
\title{Convert an annual xts object with multiple columns and to annual}
\usage{
a_to_q_xts(x, type)
}
\arguments{
\item{type}{}
}
\description{
Uses the a_to_q function.
}
