% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model_dyn.R
\name{ar_add_predictions}
\alias{ar_add_predictions}
\title{Make predictions based on growing out dlogs, after filtering to particular geography.}
\usage{
ar_add_predictions(area_sh_use, train_start_date, train_end_date,
  test_start_date, test_end_date, model, dependentvar, depend_in_levels,
  data)
}
\arguments{
\item{data}{}
}
\description{
Make predictions based on growing out dlogs, after filtering to particular geography.
}
