% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cagr.R
\name{calc_annualized}
\alias{calc_annualized}
\title{Compounded Annual Rate of Change (not sure the difference)
(((x(t)/x(t-1)) ** (n_obs_per_yr)) - 1) * 100}
\usage{
calc_annualized(x, p)
}
\arguments{
\item{p}{}
}
\description{
Compounded Annual Rate of Change (not sure the difference)
(((x(t)/x(t-1)) ** (n_obs_per_yr)) - 1) * 100
}

