% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cagr.R
\name{calc_cagr}
\alias{calc_cagr}
\title{Compounded Annual Rate of Change
(Ending Value / Initial Value) ^ ( 1 / # of periods) - 1)}
\usage{
calc_cagr(x, n)
}
\arguments{
\item{n}{}
}
\value{

}
\description{
Compounded Annual Rate of Change
(Ending Value / Initial Value) ^ ( 1 / # of periods) - 1)
}
