% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_reg.R
\name{center_reg}
\alias{center_reg}
\title{Adds a centered regressor as a new column}
\usage{
center_reg(x, hold_var)
}
\arguments{
\item{x}{is the name of the original dataframe}

\item{hold_var}{is the character vector name of the variable for which you want to calculate a centered regressor}
}
\description{
Adds a centered regressor as a new column
}

