% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_up_trend.R
\name{clean_up_trend}
\alias{clean_up_trend}
\title{Cleans up a dataframe of data that has been read in from the Raw Data tab of a monthly trend report}
\usage{
clean_up_trend(df, seggeo, trend_type)
}
\arguments{
\item{df}{Source data}

\item{seggeo}{Character string to name the segment or geography of data}

\item{trend_type}{If "international", expects an exchange rate column to be in the source data}
}
\description{
Cleans up a dataframe of data that has been read in from the Raw Data tab of a monthly trend report
}
