% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/work_lists.R
\name{create_named_list}
\alias{create_named_list}
\title{Create a named list}
\usage{
create_named_list(vector_of_variable_names)
}
\arguments{
\item{vector_of_variable_names}{}
}
\description{
Based on idea here
http://r.789695.n4.nabble.com/how-to-make-list-return-a-list-of-named-elements-td2720620.html
}
\details{
I've used this to create a list of ts objects as part of working on the holiday regressors
}
