% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sa_from_sf.R
\name{create_sa_str}
\alias{create_sa_str}
\title{Use seasonal factors to create monthly or quarterl seasonally adjusted data}
\usage{
create_sa_str(df)
}
\arguments{
\item{df}{}
}
\description{
my thought process on this function had originally been that it would start
with a dataframe that had a single underscore between the segment code and
the variable code. So all segment and geographic info would be in the segment
code. But then I got into the Host work, and realized that it was potentially
useful to have additional segment, geography, country information in the
mneomonic, separated by underscores. But I couldn't figure out how to write a
regular express than would just break of the lodging variable concept. What I
could do was break up the other stuff until I was left with the lodging
concept. But that pointed to the idea of staying flexible on the set up of
the dataframe going into this function. So it is more generalized. Also, I
realized that the monthly and quarterly were doing the same thing, so I
combined them. I left the monthly and quarterly ones there, but I could get
rid of them once I bring the US lodfor process onto the same footing.
LAter in Sept. 2016, I wanted to add some calculations related to room changes,
but in lodfor_v2, I still seemed to be using the separate monthly and quarterly
create str steps. So I created the create_sa_str_v2 as what I would use going forward.
}

