% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sa_from_sf.R
\name{create_sa_str_q}
\alias{create_sa_str_q}
\title{use seasonal factors to create monthly seasonally adjusted data}
\usage{
create_sa_str_q(str_q)
}
\arguments{
\item{str_q}{}
}
\description{
function takes a monthly data frame with monthly data and seasonal factors
and creates monthly sa
}
