% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sa_from_sf.R
\name{create_sa_str_v2}
\alias{create_sa_str_v2}
\title{Use seasonal factors to create monthly or quarterl seasonally adjusted data
Initially created in Sept. 2016}
\usage{
create_sa_str_v2(df, freq)
}
\arguments{
\item{freq}{Character to indicate frequency, either "q", or "m"}
}
\description{
Use seasonal factors to create monthly or quarterl seasonally adjusted data
Initially created in Sept. 2016
}
