% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_dyn.R
\name{create_test_train}
\alias{create_test_train}
\title{create the test and train data}
\usage{
create_test_train(df, varlist, dates, testNA1, testNA2)
}
\arguments{
\item{testNA2}{}
}
\description{
create the test and train data
}

