% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_to_do_vec.R
\name{create_to_do_vec}
\alias{create_to_do_vec}
\title{Create a vector of geographies to be run for various equations.}
\usage{
create_to_do_vec(in_area_sh, ex_area_sh, market_vec)
}
\arguments{
\item{market_vec}{}
}
\description{
Create a vector of geographies to be run for various equations.
}
