% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_trend_tidy.R
\name{create_trend_tidy}
\alias{create_trend_tidy}
\title{Append a trend variable}
\usage{
create_trend_tidy(df, date_var, frequency)
}
\arguments{
\item{df}{Data frame to be modified}

\item{date_var}{Column containing the date (enter as character vector)}

\item{frequency}{Number of period per year (enter as number)}
}
\value{

}
\description{
Based on a date column, append a trend variable to a data frame
}
\examples{
  create_trend(frequency=4, date_var = "date")

}
