% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_reg.R
\name{create_ts_reg_m}
\alias{create_ts_reg_m}
\title{Creates a ts object out of a selected column in a data frame with a date column
Currently works for monthly}
\usage{
create_ts_reg_m(x, var)
}
\arguments{
\item{var}{}
}
\description{
Creates a ts object out of a selected column in a data frame with a date column
Currently works for monthly
}
