% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_in.R
\name{days_in_month}
\alias{days_in_month}
\title{return days in month}
\usage{
days_in_month(d = Sys.Date(), leap_impact = 1)
}
\arguments{
\item{leap_impact}{}
}
\description{
I modified so you can set leap_impact equal 0 to help with str data
}
\examples{
days_in_month("2014-11-25")
}
