% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esttrend.R
\name{esttrend}
\alias{esttrend}
\title{Estimate a trend}
\usage{
esttrend(x, form)
}
\arguments{
\item{form}{}
}
\value{

}
\description{
Can be based on log of original series or linear. Extracts trend term
coefficient. Works on an xts object, maybe also zoo.
}
