% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_dyn.R
\name{eval_test_actual}
\alias{eval_test_actual}
\title{evaluates the test versus actual}
\usage{
eval_test_actual(full_data, test_data, dependentvar, end_date, market_text,
  test_start)
}
\arguments{
\item{end_date}{}
}
\description{
evaluates the test versus actual
}

