% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_strdays.R
\name{extend_strdays}
\alias{extend_strdays}
\title{Extend the number of strdays by recalculating}
\usage{
extend_strdays(df, freq)
}
\arguments{
\item{df}{Requires a "date" column}

\item{freq}{Either "m", "q" or "a"}
}
\description{
Extend the number of strdays by recalculating
}
\examples{
extend_strdays(freq="m")
}
