% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model_dyn.R
\name{extract_pred_lev}
\alias{extract_pred_lev}
\title{Pull out the predictation data column.}
\usage{
extract_pred_lev(pred_lev_data, depvar_lev_chr)
}
\arguments{
\item{pred_lev_data}{Just the test period data, includes predicted value
of the dependent variable in levels.}

\item{depvar_lev_chr}{name of the series to provide a bit of confirmation}
}
\value{

}
\description{
Pull out the predictation data column.
}
