% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_extend.R
\name{factor_extend}
\alias{factor_extend}
\title{Function to extract seasonal factors and extend them using a
a simple forecast. Used in seasonal_overall function.}
\usage{
factor_extend(df, horizon)
}
\arguments{
\item{horizon}{Future horizon over which to extend series, such as "5 years"}
}
\value{

}
\description{
Function to extract seasonal factors and extend them using a
a simple forecast. Used in seasonal_overall function.
}
