% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_dyn.R
\name{fit_dyn}
\alias{fit_dyn}
\title{Fits a model}
\usage{
fit_dyn(formula, data)
}
\arguments{
\item{data}{}
}
\description{
Simple approach to fit a model based on a data frame using the dyn package.
Includes a step to convert the data frame to xts before fitting.
}
