% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model_dyn.R
\name{fit_model_dyn}
\alias{fit_model_dyn}
\title{Fit a model after filtering data frame to particular geography and traning period.}
\usage{
fit_model_dyn(area_sh_use, train_start_date, train_end_date, spec, data)
}
\arguments{
\item{data}{}
}
\description{
Fit a model after filtering data frame to particular geography and traning period.
}

