% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model_dyn.R
\name{fit_model_dyn_c}
\alias{fit_model_dyn_c}
\title{Title}
\usage{
fit_model_dyn_c(area_sh_use, train_start_date, train_end_date, spec, data)
}
\arguments{
\item{data}{}
}
\description{
Title
}
\examples{
Helpful to do as follows, putting possibly wrapper, with NA,
around the call on the function.

new_1 <- mkeq_use_3 \%>\%
mutate(
 model = pmap(.l=list(area_sh, train_start_date,
                      train_end_date, spec),
              .f=possibly(arlodr::fit_model_dyn_c, NA), data = df_use1)
)

}
