% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred.R
\name{getFredTable_list_aranmod}
\alias{getFredTable_list_aranmod}
\title{Get a Fred table list}
\usage{
getFredTable_list_aranmod(search_tags1, search_tags2, freq_var, drop_title,
  keep_title, series_id, series_id_gsubend, ...)
}
\arguments{
\item{...}{}
}
\description{
Get a Fred table list
}

