% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_set_scale.R
\name{ggplot_set_scale_y_con}
\alias{ggplot_set_scale_y_con}
\title{set format of scale_y_continuous}
\usage{
ggplot_set_scale_y_con(plot1, set_axis_type)
}
\arguments{
\item{set_axis_type}{either "dollar" or "percent"}
}
\description{
Created to be used in automated creation of graphs when the format of the y-axis (horizontal
axis when used in bar graph after coord_flip()) needs to be set based on a variable
(e.g. graphing concepts such as occupancy and revpar in the same process.
}
\examples{




}
