% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_dlog.R
\name{grow_dlog}
\alias{grow_dlog}
\title{take a test dataframe with predicted series of dlogs and use it to grow out the actual series of interest
Also see grow_dlog2.}
\usage{
grow_dlog(predict_xts, train_xts, predict_dlog, predict_level)
}
\arguments{
\item{predict_level}{}
}
\description{
take a test dataframe with predicted series of dlogs and use it to grow out the actual series of interest
Also see grow_dlog2.
}

