% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_dlog.R
\name{grow_dlog2}
\alias{grow_dlog2}
\title{Take a test data frame with predicted series of dlogs and use it to grow out the actual series of interest.
Similar to grow_dlog, but this uses data in df format, while the other uses data in xts format. Also, this
returns the data in df format.}
\usage{
grow_dlog2(predict_df, train_df, predict_dlog, predict_level)
}
\arguments{
\item{predict_level}{}
}
\description{
Take a test data frame with predicted series of dlogs and use it to grow out the actual series of interest.
Similar to grow_dlog, but this uses data in df format, while the other uses data in xts format. Also, this
returns the data in df format.
}

