% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{holiday_reg}
\alias{holiday_reg}
\title{Holiday regressors}
\format{

}
\usage{
holiday_reg
}
\description{
Monthly dummary variables indicating holidays
}
\details{
Temporary set up to contain the monthly holiday regressors I had set up at one point.
}
\keyword{datasets}
