% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_index.R
\name{index_m}
\alias{index_m}
\title{create an index of a _monthly_ xts or maybe zoo series}
\usage{
index_m(x, index_year)
}
\arguments{
\item{index_year}{}
}
\description{
create an index of a _monthly_ xts or maybe zoo series
}
