% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_index.R
\name{index_m_melted}
\alias{index_m_melted}
\title{create an index of a single monthly series that is in a melted dataframe}
\usage{
index_m_melted(x, index_year)
}
\arguments{
\item{index_year}{}
}
\value{

}
\description{
create an index of a single monthly series that is in a melted dataframe
}
