% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_index.R
\name{index_q}
\alias{index_q}
\title{create an index of a quarterly xts or maybe zoo series}
\usage{
index_q(x, index_year)
}
\arguments{
\item{index_year}{}
}
\value{

}
\description{
create an index of a quarterly xts or maybe zoo series
}
