% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_index.R
\name{index_q_melted}
\alias{index_q_melted}
\title{create an index of a single quarterly series that is in a melted dataframe}
\usage{
index_q_melted(x, index_year)
}
\arguments{
\item{index_year}{}
}
\value{

}
\description{
create an index of a single quarterly series that is in a melted dataframe
}
