% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_index.R
\name{index_q_xts}
\alias{index_q_xts}
\title{runs across a quarterly xts object to create an index of each series}
\usage{
index_q_xts(x, index_year)
}
\arguments{
\item{index_year}{}
}
\description{
runs across a quarterly xts object to create an index of each series
}
