% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_str.R
\name{load_str}
\alias{load_str}
\title{process monthly STR data}
\usage{
load_str(load_m)
}
\arguments{
\item{load_m}{}
}
\description{
function takes a data frame of monthly str data
and returns a list containing a monthly data frame
and a quarterly data frame
}

