% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_str.R
\name{load_str_m}
\alias{load_str_m}
\title{Process monthly STR data}
\usage{
load_str_m(load_m, units = NULL)
}
\arguments{
\item{load_m}{Takes a wide dataframe with monthly lodging date as an input (e.g. with
each series in a column).}

\item{units}{"millions", "thousands" or "ones" Will be used to divide
supt, demt, rmrevt, supd, and demd.}
}
\value{
Returns a wide dataframe with the processed
STR data. It will contain:
adr
days
demd
demt
occ
revpar
rmrevt
supd
supt
}
\description{
Previously this was the first part of load_str, which had a second part that
did the quarterly steps. I've broken it apart in July 2016 so that I could
make tweaks to the quarterly part, and it seemed unnecessary to keep all
together as one step.
}

