% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_str.R
\name{load_str_q}
\alias{load_str_q}
\title{Process monthly STR data to create quarterly data}
\usage{
load_str_q(load_m)
}
\arguments{
\item{load_m}{Monthly STR data, after it is processed in load_str_m}
}
\value{
Dataframe with quarterly STR data. In addition to the series in the
monthly output, which are:
adr
days
demd
demt
occ
revpar
rmrevt
supd
supt
It will also contain:
sups (start-of-period supply, based on the first month of the quarter)
supe (end-of-period supply, based on first month of following quarter)
supem (ending month supply, supply in last month of quarter)
}
\description{
This was originally the second part of the load_str step. I broke that
apart. I also added to this so that it creates some end-of-period,
and start-of-period supply data.
}
