% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_beta_log.R
\name{log_growth_1}
\alias{log_growth_1}
\title{Estimate annualized growth}
\usage{
log_growth_1(df, periods_year)
}
\arguments{
\item{periods_year}{the number of periods in a year, e.g. 4 for quarters or
365.25 for days. This will convert the slope coefficient to an annualized growth
rate by doing ((1 + estimate) ^ periods_year - 1)}
}
\value{

}
\description{
See example_growth_trend in vignettes for usage. This makes an estimate even
if there are only a couple actual data points.
}
