% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_beta_log.R
\name{log_growth_2}
\alias{log_growth_2}
\title{Estimate annualize growth modified}
\usage{
log_growth_2(df, periods_year, y, x)
}
\arguments{
\item{periods_year}{the number of periods in a year, e.g. 4 for quarters or
365.25 for days. This will convert the slope coefficient to an annualized growth
rate by doing ((1 + estimate) ^ periods_year - 1)}

\item{y}{character vector naming dependent variable}

\item{x}{character vector naming independent variable}
}
\description{
See example_growth_trend in vignettes for usage. This makes an estimate even
if there are only a couple actual data points.
}
\details{
Same as log_growth_1, but modified to allow refering to columns by name
as arguments, rather than requiring that they be named x and y
}
