% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_plot.R
\name{plot_title_2}
\alias{plot_title_2}
\title{another approach to add title}
\usage{
plot_title_2(plot, grtitle, footnote, filename)
}
\arguments{
\item{filename}{}
}
\description{
another approach to add title
}
