% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_plot.R
\name{plot_title_3}
\alias{plot_title_3}
\title{another approach to add a title}
\usage{
plot_title_3(plot, grtitle, footnote, filename)
}
\arguments{
\item{filename}{}
}
\description{
in the following I had added ?filename? in the fred version.
I included in the combined document. Not sure what impact that will have
}
