% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_plot.R
\name{plot_title_5}
\alias{plot_title_5}
\title{Plotting function}
\usage{
plot_title_5(
  plot,
  grtitle,
  subtitle,
  footnote,
  filename,
  saveheight = 5.7,
  savewidth = 8.5
)
}
\arguments{
\item{savewidth}{}
}
\value{

}
\description{
This is the plotting function that I was using for some time in the US overview graphs. I had previously had some
troubling moving it to the package. But when I cleaned up how I was referring to paths, it seemed
to work. So here it is.
}
