% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_plot.R
\name{plot_title_two1}
\alias{plot_title_two1}
\title{combines two plots together (e.g. employment and GDP)}
\usage{
plot_title_two1(p1, p2, grtitle, footnote, filename)
}
\arguments{
\item{filename}{}
}
\description{
to get this function to work I needed to modify ggsave using the following line
this is because I've used ggplot_gtable, and so the pieces aren't ggplot plot
elements, and that's what's expected in ggsave. Based on:
http://stackoverflow.com/questions/18406991/saving-a-graph-with-ggsave-after-using-ggplot-build-and-ggplot-gtable
this should no longer be necessary, so I've commented it out
ggsave <- ggplot2::ggsave; body(ggsave) <- body(ggplot2::ggsave)[-2]
}
