% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_dyn.R
\name{predict_dyn}
\alias{predict_dyn}
\title{create predictions using dyn$predict
created based on
http://stackoverflow.com/a/13098860/1770086
also based on the help for dyn, see example at end}
\usage{
predict_dyn(model, train, test, dependentvar)
}
\arguments{
\item{dependentvar}{}
}
\description{
create predictions using dyn$predict
created based on
http://stackoverflow.com/a/13098860/1770086
also based on the help for dyn, see example at end
}

