% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_dyn.R
\name{predict_dyn2}
\alias{predict_dyn2}
\title{Create predictions using dyn$predict, setup to start with and return
data frames rather than xts}
\usage{
predict_dyn2(model, train, test, dependentvar)
}
\arguments{
\item{model}{Typically model fitted with dyn$lm}

\item{train}{Data frame with first column in date format so that function
can convert to xts easily for calcs.}

\item{test}{Data frame with first column in date format.}

\item{dependentvar}{Character string naming the dependentvar that is to be
grown out iteratively}
}
\value{

}
\description{
In other words, in this version, you can supply data frames as arguments
and receive a data frame as output. This is a bit friendlier with a map
approach.
}
\details{
created based on
http://stackoverflow.com/a/13098860/1770086
also based on the help for dyn, see example at end
}
