% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred.R
\name{pull_emp}
\alias{pull_emp}
\title{pull data from FRED}
\usage{
pull_emp(series_id, ac_substr_1, ac_substr_2, var_substr_1, var_substr_2, freq,
  ...)
}
\arguments{
\item{...}{}
}
\description{
pull data from FRED
}

