% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred.R
\name{pull_list}
\alias{pull_list}
\title{apply a function given a list of inputs, extract results}
\usage{
pull_list(pull_fun, list_input)
}
\arguments{
\item{list_input}{}
}
\description{
apply a function given a list of inputs, extract results
}

