% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_conversion.R
\name{q_to_m}
\alias{q_to_m}
\title{Conversion from quarterly to monthly}
\usage{
q_to_m(x, type)
}
\arguments{
\item{type}{}
}
\description{
Follows similar approach as annual to quarterly and could be collapsed as as a single function
}
