% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_conversion.R
\name{q_to_m_xts}
\alias{q_to_m_xts}
\title{Convert a quarterly xts object with multiple columns to monthly}
\usage{
q_to_m_xts(x, type)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Uses the q_to_m function.
}
