% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind.R
\name{rbind.all.columns}
\alias{rbind.all.columns}
\title{rbind dataframes when you want to keep all columns}
\usage{
rbind.all.columns(x, y)
}
\arguments{
\item{y}{}
}
\description{
simple function to
of both data frames
from: https://amywhiteheadresearch.wordpress.com/2013/05/13/combining-dataframes-when-the-columns-dont-match/
}

