% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoder_func.R
\name{recoder_func}
\alias{recoder_func}
\title{Recoder}
\usage{
recoder_func(data, oldvalue, newvalue)
}
\arguments{
\item{newvalue}{}
}
\value{

}
\description{
Sets up a way to recode a variable, similar to a lookup table approach
copied from following link
http://susanejohnston.wordpress.com/2012/10/01/find-and-replace-in-
r-part-2-how-to-recode-many-values-simultaneously/
}
