% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_beta_log.R
\name{roll_beta_log}
\alias{roll_beta_log}
\title{Calculates rolling regression estimates}
\usage{
roll_beta_log(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Uses rollapply to estimate equation results.
}
