% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_indiv_r.R
\name{run_indiv_r}
\alias{run_indiv_r}
\title{run individual regressions for each metro}
\usage{
run_indiv_r(
  df,
  eq_id,
  start_date,
  end_date,
  gp_vec,
  y,
  x1,
  x2,
  x3,
  x4,
  x5,
  use_ar,
  use_c,
  drop_area_sh,
  select_out
)
}
\arguments{
\item{select_out}{}
}
\value{

}
\description{
original part of regress script in fred project
}
