% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pooled.R
\name{run_pooled_r}
\alias{run_pooled_r}
\title{run pooled and fixed effect regressions}
\usage{
run_pooled_r(
  df,
  eq_id,
  start_date,
  end_date,
  gp_vec,
  y,
  x1,
  x2,
  x3,
  x4,
  x5,
  use_ar,
  use_c,
  drop_area_sh,
  select_out
)
}
\arguments{
\item{select_out}{}
}
\description{
run pooled and fixed effect regressions
}
