% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_percent.R
\name{scale_percent_1}
\alias{scale_percent_1}
\title{Set scale to percent with one decimal place}
\usage{
scale_percent_1(x)
}
\arguments{
\item{accuracy}{integer to determine the number of places to show after decimal point}
}
\description{
Set scale to percent with one decimal place
}
\examples{
In a ggplot plot, use the following to set to y-axis in percent with one decimals.
scale_y_continuous(labels = arlodr::scale_percent_1, limits=c(.02,NA))

}
