% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_percent.R
\name{scale_percent_2}
\alias{scale_percent_2}
\title{Set scale to percent with two decimal places}
\usage{
scale_percent_2(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Set scale to percent with two decimal places
}
\examples{
#' In a ggplot plot, use the following to set to y-axis in percent with two decimals.
#' scale_y_continuous(labels = arlodr::scale_percent_2, limits=c(.02,NA))
}
