% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_seas_factors.R
\name{seas_factors_m}
\alias{seas_factors_m}
\title{create monthly seasonal factors}
\usage{
seas_factors_m(str_m, dont_m_cols)
}
\arguments{
\item{dont_m_cols}{}
}
\description{
function takes a data frame of monthly str data and returns a set of monthly
seasonal factors, including future months
}

