% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_seas_factors.R
\name{seas_factors_m_2}
\alias{seas_factors_m_2}
\title{modifed version of seas_factors_m}
\usage{
seas_factors_m_2(str_m, dont_m_cols, hold_reg)
}
\arguments{
\item{hold_reg}{}
}
\description{
modifed version of seas_factors_m
}

