% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_seas_factors.R
\name{seas_factors_q}
\alias{seas_factors_q}
\title{create quarterly seasonal factors}
\usage{
seas_factors_q(str_q, dont_q_cols)
}
\arguments{
\item{dont_q_cols}{}
}
\description{
function takes a data frame of quarterly str data and returns a set of
quarterly seasonal factors, including future months
}

