% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arseas.R
\name{seasonal_ad}
\alias{seasonal_ad}
\title{seasonal_ad}
\usage{
seasonal_ad(x, meffects = c("const", "easter[8]", "thank[5]"),
  qeffects = c("const", "easter[8]", "tdnolpyear"))
}
\arguments{
\item{qeffects}{}
}
\description{
seasonally adjust a series as aruments series output a dataframe containing
the seasonal factor and the seasonally adjusted series
}

