% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arseas.R
\name{seasonal_ad_2}
\alias{seasonal_ad_2}
\title{modified version of seasonal_ad
Uses a holiday regressor argument,
also set up to use seats, following the way I had set things up in my more structured approach}
\usage{
seasonal_ad_2(x, meffects = c("tdnolpyear", "ls2001.Sep"),
  qeffects = c("tdnolpyear", "ls2001.Sep"), hold_reg)
}
\arguments{
\item{hold_reg}{}
}
\description{
modified version of seasonal_ad
Uses a holiday regressor argument,
also set up to use seats, following the way I had set things up in my more structured approach
}

