% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_apply.R
\name{seasonal_apply}
\alias{seasonal_apply}
\title{Apply the seasonal factors}
\usage{
seasonal_apply(df, date, group)
}
\arguments{
\item{df}{Data frame in a tidy format with seasonal factors already joined on}

\item{date}{Name of the data column (used in organizing columns for output)}

\item{group}{Name of geography or other grouping, e.g., geoseg.}
}
\value{

}
\description{
Apply the seasonal factors
}
\examples{
m_us_1 <- raw_m_us \%>\%
}
