% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_est.R
\name{seasonal_est}
\alias{seasonal_est}
\title{Estimate seasonal factors}
\usage{
seasonal_est(df, holiday_input = "blank")
}
\arguments{
\item{df}{Needs to have a date column, a group column that can be}

\item{holiday_input}{Optional set of monthly holiday regressors as a
time series}
}
\value{

}
\description{
Estimate seasonal factors
}
