% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_overall.R
\name{seasonal_overall}
\alias{seasonal_overall}
\title{Combined steps of seasonal adjustment}
\usage{
seasonal_overall(df, group, holiday_input = "blank", horizon = "5 years")
}
\arguments{
\item{df}{should be a data frame that contains a date column, a group column, a var column, and a value column. var is the column that contains the variable names, such as var.}

\item{group}{grouping column, such as geoseg. Entered without quotes.}

\item{horizon}{}
}
\value{

}
\description{
Combined steps of seasonal adjustment
}
\examples{
output_m <- seasonal_overall(work_m, group=geoseg,
holiday_input=hold_reg, horizon="5 years")
}
