% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_xts.R
\name{simp_xts}
\alias{simp_xts}
\title{put a tidy data frame with identfying columns
such as a geographic column and a date column into xts format}
\usage{
simp_xts(x, to_exclude)
}
\arguments{
\item{to_exclude}{}
}
\description{
x is a tidy dataframe with variables as columns
to_exclude a character vector of columns to exclude in the gather step
such as c("date", "area_sh")
creates a character vector of columns to exclude in the gather step
I ended up not going this route, and instead requiring the character
vector as an argument, but this is still a useful example of what's
possible
to_exclude <- c("date", interp(geo_code))
}
\details{
Still requires that one of the columns be titled area_sh. I guess I left that part hard coded.

based on example from:
http://stackoverflow.com/questions/28911496/exclusions-with-when-using-string-versions-underscore-suffix-such-as-gather

when working with BLS data I had to add the following step
before read.zoo because I was getting an index error
   as.data.frame() %>%
}

