% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arseas.R
\name{skip_seasonal_ad}
\alias{skip_seasonal_ad}
\title{skip seasonal adjustment}
\usage{
skip_seasonal_ad(x)
}
\arguments{
\item{x}{}
}
\description{
skip seasonal adjustment but still output series that are the same format as
what would be exported by the seasonal adjustment function in other words
copy the unadjusted series as the seasonally adjusted, create seasonal
factors equal to 1 and create a temporary fct series
}
