% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_themes.R
\name{theme_ts1}
\alias{theme_ts1}
\title{defined theme for use in creating plots}
\usage{
theme_ts1(base_size = 14, base_family = "Arial")
}
\arguments{
\item{base_family}{}
}
\value{

}
\description{
defined theme for use in creating plots
}
