% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_millions.R
\name{units_millions}
\alias{units_millions}
\title{units_millions}
\usage{
units_millions(col)
}
\arguments{
\item{col}{}
}
\description{
converts units to millions, taking a column of a dataframe as input
}

