% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esttrend.R
\name{xts_dropallna}
\alias{xts_dropallna}
\title{Drop columns that are all na}
\usage{
xts_dropallna(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Works on an xts object. Useful to run before a rollapply regression.
}
