% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model_dyn.R
\name{extract_pred_lev}
\alias{extract_pred_lev}
\title{Pull out the predictation data column.}
\usage{
extract_pred_lev(pred_lev_data, depvar_lev_chr)
}
\arguments{
\item{depvar_lev_chr}{}
}
\description{
Pull out the predictation data column.
}

